+++ tools/goctl/util/readlink+polyfill.go
@@ -1,0 +1,8 @@
+//go:build windows
+// +build windows
+
+package util
+
+func ReadLink(name string) (string, error) {
+	return name, nil
+}
